requirements.txt
torch
faster-whisper
sentence-transformers
chromadb
streamlit
pydub
numpy
tqdm
rich
openai
python-dotenv
# Optional (comment out if you don't need them)
# pyannote.audio
# rank-bm25
# transformers
# openai
